requirements.txt
aiohttp>=3.9
aiofiles>=23.0
feedparser>=6.0
beautifulsoup4>=4.12
rich>=13.0
textual>=0.40.0
python-dateutil>=2.8
trafilatura>=1.8
structlog>=24.0
pydantic>=2.0
python-dotenv>=1.0

pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=4.1
pytest-aiohttp>=1.0
ruff>=0.1
mypy>=1.8
